src/knowledge-service/requirements.txt
llama-index>=0.12.0
llama-index-embeddings-dashscope>=0.3.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.1.0
